requirements.txt
django==2.2.17
python-ldap==3.1.0
